Clean up system specs
