+++ spec/system/effort_show/visit_spec.rb
