+++ spec/system/event_group_roster/manage_check_in_spec.rb
