+++ spec/system/event_group_roster/start_ready_efforts_spec.rb
