+++ spec/system/event_group_roster/visit_spec.rb
